Cloneable client (#9)
* create helper function for getting the recent tracks page

* implement `Clone` trait for `Client`

This is done by replacing `Box<dyn RetryStrategy>` with `Arc<dyn RetryStrategy>`.
You *could* also accomplish this by making `Box<dyn RetryStrategy>` cloneable with a crate like `dyn-clone`,
but most of the APIs for the `RetryStrategy` trait suggest that you'd never want to store state in a retry strategy,
so we don't need to transfer ownership using a `Box`. Also, implementations of `RetryStrategy` would also have to implement `Clone`.